Add Kaggle API automation and fix env/fidelity issues found running it
- configs/phi3_cossim.yaml: dtype auto -> float32 (fidelity-matched
  default was never actually checked in)
- trust_remote_code: true -> false (Phi-3's remote modeling code breaks
  under newer transformers' rope_scaling handling; use the built-in
  implementation instead -- logged as a fidelity deviation)
- cos_sim.py: cast to float32 before .numpy() (bfloat16 isn't numpy-native)
- kaggle/, scripts/run_on_kaggle.sh: push/poll/pull automation for running
  the reproduction on Kaggle's free GPU via the API instead of the
  notebook UI, including a P100/sm_60 compatibility workaround for
  Kaggle's GPU-assignment bug

All deviations logged in docs/KNOWN_DISCREPANCIES.md (#7-#9).

+++ src/safety_layers_repro/cos_sim.py
@@ -149,7 +149,10 @@ def get_last_position_hidden_states(
     for i in range(len(step0_layers)):
         if i == 0:
             continue  # skip embedding layer, matches original script
-        vectors.append(step0_layers[i][0][-1].detach().cpu().numpy())
+        # .numpy() doesn't support bfloat16 directly (e.g. under dtype=auto,
+        # which resolves to this model's default torch_dtype); cast to
+        # float32 first. No-op for configs already running in float32.
+        vectors.append(step0_layers[i][0][-1].detach().to(torch.float32).cpu().numpy())
     return vectors
